Added skeleton for main functionality.
The class Generator is responsible for the generation of PlantUML code based on the PHP code from input. The basic fromtString() method takes a string of PHP code and returns a string of PlantUML code.

Added a test with the most basic conversion I could think of. It expects perfect output format. Probably will have to do som basic pattern matching at some point. Syntax checking would be useful, but I'll have to research that.