Proof of concept.
Added namespace to test.
Moved PhpParser to be an external dependency of Generator to allow the caller to decide the options. Unsure about this as Generator could easily be just take options instead.
Implemented a very basic proof-of-concept logic that passes the test. Want to refactor the result generation to use a template engine as well as refine the actual translation.